+++ calculator-api/services/calculation/calculation.go
@@ -5,5 +5,5 @@ type ICalculationService interface {
 	Substract(number1, number2 float64) (float64, error)
 	Multiply(number1, number2 float64) (float64, error)
 	Divide(dividend, divisor float64) (float64, error)
-	Sum(numbers []int) (int, error)
+	Sum(numbers []float64) (float64, error)
 }
